Move tag logic from deleted view-file-page.js to file-page.js
